ISD-192 test git hubs container registry performance for existing workflows (#146)
* ci(ISD-192): Push images to ghcr.io

This change replaces the CI behavior of building a docker image locally and passing it around as an artifact to other CI steps with using Githubs' container registry. The approach simplifies the CI (as microk8s now only authenticates towards the registry and pull the images, without needing to download artifacts, build a local registry and push/pull the images from there), it also shaves off a couple of minutes (5-6) from the whole CI scenario.