ISD-203 indico export uwsgi metrics to prometheus (#153)
feat(ISD-203): Use statsd_exporter for uWSGI metrics

This feature adds the capability of pushing metrics to a statsd server to the uWSGI component of Indico.
For this, uWSGI is installed with the statsd plugin enabled and configured to push to the default port of
statsd-prometheus-exporter, which is now part of the charm. The received statsd metrics are exposed
as prometheus metrics to be scraped at a /metrics endpoint under a (statsd exporter default) port.
The changes are reflected in unit and integration tests.